requirements.txt
Flask~=1.1.2
SQLAlchemy~=1.3.19
Werkzeug~=1.0.1
python-dotenv~=0.19.1
requests~=2.24.0
gunicorn~=20.1.0
Flask-Bootstrap~=3.3.7.1
Flask-CKEditor~=0.4.4.1
Flask-Gravatar~=0.5.0
Flask-Login~=0.5.0
Flask-SQLAlchemy~=2.4.4
Flask-WTF~=0.14.3
